The last version

+++ DebtManagement/User/ctrlPageUser.cs
@@ -11,7 +11,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using DevExpress.XtraGrid.Views.Grid;
 using InternalDialog = DevExpress.Utils.CommonDialogs.Internal;
+using System.Windows.Controls;
+using DevExpress.XtraMap;
+using DevExpress.XtraReports.UI;
 namespace DebtManagement
 {
     public partial class ctrlPageUser : DevExpress.XtraEditors.XtraUserControl
@@ -49,13 +53,20 @@ namespace DebtManagement
         {
             // Select Current UserID.
             int UserID = -1;
+            string UserType = string.Empty;
 
             if (gridViewUsers != null && gridViewUsers.FocusedRowHandle >= 0)
             {
                 UserID = Convert.ToInt32(gridViewUsers.GetRowCellValue(gridViewUsers.FocusedRowHandle, "UserID"));
+                UserType = gridViewUsers.GetRowCellValue(gridViewUsers.FocusedRowHandle, "UserType").ToString();
             }
 
-            frmAddUpdateUser frm = new frmAddUpdateUser(UserID);
+            // If There Are Only One Manager Should Disable Select User Type (In Form "frmAddUpdateUser").
+            bool isThereOnlyOnManager = isThereOnlyOneManagerInGrid();
+
+            bool EnableSelectUserType = (UserType == "مدير" && isThereOnlyOnManager) ? false : true;
+
+            frmAddUpdateUser frm = new frmAddUpdateUser(UserID, EnableSelectUserType);
 
             frm.ShowDialog();
 
@@ -64,30 +75,72 @@ namespace DebtManagement
 
         private async void btnDelete_Click(object sender, EventArgs e)
         {
+
             // Select Current UserID.
             int UserID = -1;
             string UserName = "";
+            string UserType = "";
+            bool isThereOnlyOneManager = false;
+
 
             if (gridViewUsers != null && gridViewUsers.FocusedRowHandle >= 0)
             {
                 UserID = Convert.ToInt32(gridViewUsers.GetRowCellValue(gridViewUsers.FocusedRowHandle, "UserID"));
                 UserName = gridViewUsers.GetRowCellValue(gridViewUsers.FocusedRowHandle, "UserName").ToString();
+                UserType = gridViewUsers.GetRowCellValue(gridViewUsers.FocusedRowHandle, "UserType").ToString();
+
+                isThereOnlyOneManager = isThereOnlyOneManagerInGrid();
             }
 
-            var result = frmMessageBoxDev.ShowDialog($"تاكيد عملية حذف المستخدم [{UserName}]", "تحذير", "حذف", "الغاء", frmMessageBoxDev.ModeDialog.Question);
+
+            // If There Are Only One Manager Don't Deleted (Out Of this Function).
+            if (isThereOnlyOneManager && UserType == "مدير")
+            {
+                frmMessageBoxDev.ShowDialog("لاتستطيع حذف هذا الحساب لانه المدير الوحيد", "خطا", "حسنا", frmMessageBoxDev.ModeDialog.Error, frmMessageBoxDev.Focus.btn1);
+                return;
+            }
+
+
+            var result = frmMessageBoxDev.ShowDialog($"تاكيد عملية حذف {UserType} [{UserName}]", "تحذير", "حذف", "الغاء", frmMessageBoxDev.ModeDialog.Question);
 
             if (result == InternalDialog.DialogResult.Yes)
             {
                 if (await clsUser.DeleteUserAsync(UserID))
                 {
                     await LoadUser();
-                    frmMessageBoxDev.ShowDialog("تم حذف المستخدم", "نجت العملية", "حسنا", frmMessageBoxDev.ModeDialog.Information, frmMessageBoxDev.Focus.btn1);
+                    frmMessageBoxDev.ShowDialog($"تم حذف {UserType}", "نجت العملية", "حسنا", frmMessageBoxDev.ModeDialog.Information, frmMessageBoxDev.Focus.btn1);
                 }
                 else
                 {
-                    frmMessageBoxDev.ShowDialog("حدث خطا اثناء حذف المستخدم", "حدث خطا", "حسنا", frmMessageBoxDev.ModeDialog.Error, frmMessageBoxDev.Focus.btn1);
+                    frmMessageBoxDev.ShowDialog($"حدث خطا اثناء حذف {UserType}", "حدث خطا", "حسنا", frmMessageBoxDev.ModeDialog.Error, frmMessageBoxDev.Focus.btn1);
                 }
             }
+        }
+
+        
+        private bool isThereOnlyOneManagerInGrid()
+        {
+            string FiledName = "UserType";
+            string targetValue = "مدير";
+
+            // Initialize the counter
+            int count = 0;
+
+            // Loop through all rows in the GridView
+            for (int i = 0; i < gridViewUsers.RowCount; i++)
+            {
+                // Get the value of the cell in the specified column
+                object cellValue = gridViewUsers.GetRowCellValue(i, FiledName);
+                if (cellValue != null && cellValue.ToString() == targetValue)
+                {
+                    count++;
+
+                    if (count > 1)
+                        return false;
+                }
+            }
+
+            return true;
         }
     }
 }
